--- src/components/ErrorScreen/styles.ts
@@ -1,22 +1,0 @@
-import styled from 'styled-components/native';
-
-export const Container = styled.SafeAreaView`
-  background: #fff;
-  padding: 20px;
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const Title = styled.Text`
-  color: #333;
-  font-size: 25px;
-  font-weight: bold;
-  margin-bottom: 20px;
-  text-align: center;
-`;
-
-export const Error = styled.Text`
-  color: #999;
-  font-size: 14px;
-`;
