--- src/components/LoadingScreen/styles.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components/native';
-
-export const Container = styled.SafeAreaView`
-  background: #fff;
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-`;
