+++ src/modules/LoadingScreen.tsx
@@ -1,12 +1,22 @@
 import React, { useEffect } from 'react';
 import { Animated } from 'react-native';
+import styled from 'styled-components/native';
 
 //@ts-ignore
-import logo from '../../assets/logo.png';
+import logo from '../assets/logo.png';
 
-import { Container } from './styles';
+//  ### STYLES
+
+const Container = styled.SafeAreaView`
+  background: #fff;
+  flex: 1;
+  justify-content: center;
+  align-items: center;
+`;
 
 const offset = new Animated.ValueXY({ x: 0, y: 0 });
+
+//  ### JSX
 
 export default function LoadingScreen() {
   function runAnimation() {
