--- src/pages/Create/styles.ts
@@ -1,32 +1,0 @@
-import styled from 'styled-components/native';
-
-import Input from '../../components/Input';
-import Button from '../../components/Button';
-
-export const Container = styled.View`
-  flex: 1;
-  padding: 20px;
-`;
-
-export const Title = styled.Text`
-  font-size: 25px;
-  align-self: flex-start;
-  font-weight: bold;
-  color: #333;
-`;
-
-export const Form = styled.ScrollView.attrs({
-  showsVerticalScrollIndicator: false,
-})`
-  align-self: stretch;
-  margin-top: 25px;
-  padding: 10px;
-`;
-
-export const TInput: any = styled(Input)`
-  margin-bottom: 20px;
-`;
-
-export const SubmitButton = styled(Button)`
-  margin-top: 20px;
-`;
