--- src/pages/Details/styles.ts
@@ -1,31 +1,0 @@
-import styled from 'styled-components/native';
-
-export const Container = styled.View`
-  flex: 1;
-  padding: 20px;
-  margin-top: 50px;
-`;
-
-export const Title = styled.Text`
-  color: #333;
-  font-size: 25px;
-  font-weight: bold;
-`;
-
-export const Info = styled.View`
-  margin-top: 10px;
-
-  flex-direction: row;
-  align-items: center;
-`;
-
-export const Statement = styled.Text`
-  color: #666;
-  font-size: 18px;
-`;
-
-export const Data = styled.Text`
-  color: #333;
-  font-size: 16px;
-  font-weight: bold;
-`;
